+++ src/framework/ui/internal/uiconfiguration.h
@@ -67,10 +67,10 @@ public:
     int musicalFontSize() const override;
     async::Notification musicalFontChanged() const override;
 
-    float guiScaling() const override;
-    float physicalDotsPerInch() const override;
+    double guiScaling() const override;
+    double physicalDotsPerInch() const override;
 
-    void setPhysicalDotsPerInch(std::optional<float> dpi) override;
+    void setPhysicalDotsPerInch(std::optional<double> dpi) override;
 
     QByteArray pageState(const QString& pageName) const override;
     void setPageState(const QString& pageName, const QByteArray& state) override;
@@ -113,7 +113,7 @@ private:
 
     ThemeList m_themes;
     size_t m_currentThemeIndex = 0;
-    std::optional<float> m_customDPI;
+    std::optional<double> m_customDPI;
 };
 }
